+++ src/ContaoManager/Plugin.php
@@ -8,13 +8,15 @@ use Contao\CoreBundle\ContaoCoreBundle;
 use Contao\ManagerPlugin\Bundle\BundlePluginInterface;
 use Contao\ManagerPlugin\Bundle\Config\BundleConfig;
 use Contao\ManagerPlugin\Bundle\Parser\ParserInterface;
+use Contao\ManagerPlugin\Config\ConfigPluginInterface;
 use ContaoBootstrap\Core\ContaoBootstrapCoreBundle;
 use ContaoBootstrap\Layout\ContaoBootstrapLayoutBundle;
+use Symfony\Component\Config\Loader\LoaderInterface;
 
-final class Plugin implements BundlePluginInterface
+final class Plugin implements BundlePluginInterface, ConfigPluginInterface
 {
     /**
-     * {@inheritdoc}
+     * {@inheritDoc}
      */
     public function getBundles(ParserInterface $parser): array
     {
@@ -23,9 +25,17 @@ final class Plugin implements BundlePluginInterface
                 [
                     ContaoCoreBundle::class,
                     ContaoBootstrapCoreBundle::class,
-                ]
+                ],
             );
 
         return [$bundleConfig];
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public function registerContainerConfiguration(LoaderInterface $loader, array $managerConfig): void
+    {
+        $loader->load(__DIR__ . '/../Resources/config/contao_bootstrap.yaml');
     }
 }