+++ src/Listener/Dca/LayoutDcaListener.php
@@ -19,41 +19,34 @@ use function serialize;
 
 final class LayoutDcaListener extends AbstractListener
 {
-    /** @var string */
-    // phpcs:ignore SlevomatCodingStandard.TypeHints.PropertyTypeHint.MissingNativeTypeHint
-    protected static $name = 'tl_layout';
-
-    /**
-     * Bootstrap config.
-     */
-    private Config $config;
-
-    /**
-     * Database connection.
-     */
-    private Connection $connection;
-
     /**
      * @param Config     $config     Bootstrap config.
      * @param Connection $connection Database connection.
      */
-    public function __construct(Config $config, Connection $connection, DcaManager $dcaManager)
-    {
+    public function __construct(
+        private readonly Config $config,
+        private readonly Connection $connection,
+        DcaManager $dcaManager,
+    ) {
         parent::__construct($dcaManager);
+    }
 
-        $this->config     = $config;
-        $this->connection = $connection;
+    public static function getName(): string
+    {
+        return 'tl_layout';
     }
 
     /**
      * Set the default viewport.
      *
      * @Callback(table="tl_layout", target="config.onload")
-     * @SuppressWarnings(PHPMD.Superglobals)
      */
     public function setDefaultViewPort(): void
     {
-        $GLOBALS['TL_DCA']['tl_layout']['fields']['viewport']['default'] = $this->config->get('layout.viewport', '');
+        $this->getDefinition()->set(
+            ['fields', 'viewport', 'default'],
+            $this->config->get(['layout', 'viewport'], ''),
+        );
     }
 
     /**
@@ -72,27 +65,25 @@ final class LayoutDcaListener extends AbstractListener
         }
 
         /** @psalm-var array<string,bool> $supportedFrameworkCss */
-        $supportedFrameworkCss = $this->config->get('layout.contao_framework_css', []);
+        $supportedFrameworkCss = $this->config->get(['layout', 'contao_framework_css'], []);
 
         $dataContainer->activeRecord->framework = array_values(
             array_filter(
                 StringUtil::deserialize($dataContainer->activeRecord->framework, true),
                 static fn (string $value) => $supportedFrameworkCss[$value] ?? true,
-            )
+            ),
         );
 
         $this->connection->executeStatement(
             'UPDATE tl_layout SET framework = :framework WHERE id= :id',
             [
                 'framework' => serialize($dataContainer->activeRecord->framework),
-                'id' => $dataContainer->id,
-            ]
+                'id'        => $dataContainer->id,
+            ],
         );
     }
 
-    /**
-     * @Callback(table="tl_layout", target="fields.framework.load")
-     */
+    /** @Callback(table="tl_layout", target="fields.framework.load") */
     public function frameworkOptions(string $value, DataContainer $dataContainer): string
     {
         if (! $dataContainer->activeRecord || $dataContainer->activeRecord->layoutType !== 'bootstrap') {
@@ -100,7 +91,7 @@ final class LayoutDcaListener extends AbstractListener
         }
 
         /** @psalm-var array<string,bool> $supportedFrameworkCss */
-        $supportedFrameworkCss = $this->config->get('layout.contao_framework_css', []);
+        $supportedFrameworkCss = $this->config->get(['layout', 'contao_framework_css'], []);
         $files                 = array_keys(array_filter($supportedFrameworkCss));
 
         $this->getDefinition()->set(['fields', 'framework', 'options'], $files);