+++ src/Resources/contao/languages/de/tl_layout.php
@@ -1,69 +1,42 @@
 <?php
 
-
 /**
- * fields
+ * Translations are managed automatically using Transifex. To create a new translation
+ * or to help to maintain an existing one, please register at transifex.com.
+ *
+ * Last-updated: 2022-10-12T07:10:48+00:00
  */
 
-$GLOBALS['TL_LANG']['tl_layout']['bootstrap_asseticFilter']['0']     = 'Assetic Filter';
-$GLOBALS['TL_LANG']['tl_layout']['bootstrap_asseticFilter']['1']     = 'Wählen Sie hier einen Assetic Filter oder eine Filter Chain aus, die auf diese Datei angewendet werden.';
-$GLOBALS['TL_LANG']['tl_layout']['bootstrap_conditional']['0']       = 'Conditional Comment';
-$GLOBALS['TL_LANG']['tl_layout']['bootstrap_conditional']['1']       = 'Conditional Comments ermöglichen das Einbinden Internet Explorer-spezifischer Dateien. Hier muss nur die Bedingung z.B. "lte IE 7" oder "IE 9" eingefügt werden. Der Prefix "&lt;!--[if" bzw. Sufix "]&gt;" werden automatisch eingefügt.';
-$GLOBALS['TL_LANG']['tl_layout']['bootstrap_containerClass']['0']    = 'Container-CSS-Klasse';
-$GLOBALS['TL_LANG']['tl_layout']['bootstrap_containerClass']['1']    = 'Geben Sie die CSS-Klasse an, die für den Container benutzt werden soll. Sie können beispielsweise <em>container</em>, <em>container-fluid</em> oder eine beliebige andere Klasse benutzen..';
-$GLOBALS['TL_LANG']['tl_layout']['bootstrap_containerElement']['0']  = 'Container-Element';
-$GLOBALS['TL_LANG']['tl_layout']['bootstrap_containerElement']['1']  = 'Bitte wählen Sie aus, welches Element als Bootytrap-Container dienen soll.';
-$GLOBALS['TL_LANG']['tl_layout']['bootstrap_footerClass']['0']       = 'CSS-Klasse des Fußbereichs';
-$GLOBALS['TL_LANG']['tl_layout']['bootstrap_footerClass']['1']       = 'Kann beispielsweise genutzt werden, um den Footer die Klasse .container hinzuzufügen.';
-$GLOBALS['TL_LANG']['tl_layout']['bootstrap_headerClass']['0']       = 'CSS-Klasse des Kopfbereichs';
-$GLOBALS['TL_LANG']['tl_layout']['bootstrap_headerClass']['1']       = 'Kann beispielsweise genutzt werden, um den Header die Klasse .container hinzuzufügen.';
-$GLOBALS['TL_LANG']['tl_layout']['bootstrap_importFile']['0']        = 'Datei';
-$GLOBALS['TL_LANG']['tl_layout']['bootstrap_importFile']['1']        = 'Verfügbare Datei auswählen.';
-$GLOBALS['TL_LANG']['tl_layout']['bootstrap_importJavaScripts']['0'] = 'Bootstrap JavaScripts importieren';
-$GLOBALS['TL_LANG']['tl_layout']['bootstrap_importJavaScripts']['1'] = 'Von Contao-Bootstrap zur Verfügung gestellte JavaScripts importieren.';
-$GLOBALS['TL_LANG']['tl_layout']['bootstrap_importStylesheets']['0'] = 'Bootstrap Stylesheets importieren';
-$GLOBALS['TL_LANG']['tl_layout']['bootstrap_importStylesheets']['1'] = 'Von Contao-Bootstrap zur Verfügung gestellte Stylesheets importieren.';
-$GLOBALS['TL_LANG']['tl_layout']['bootstrap_javaScripts']['0']       = 'Bootstrap JavaScripts';
-$GLOBALS['TL_LANG']['tl_layout']['bootstrap_javaScripts']['1']       = 'Bitte wählen Sie die gewünschten Dateien aus und konfigurieren sie. Sie werden beim Speichern zum Layout hinzugefügt.';
-$GLOBALS['TL_LANG']['tl_layout']['bootstrap_leftClass']['0']         = 'CSS-Grid Klassen der linken Spalte';
-$GLOBALS['TL_LANG']['tl_layout']['bootstrap_leftClass']['1']         = 'Geben Sie die CSS-Grid Klassen der linken Spalte an. Bedenken Sie, dass das Holygrail-Layout von Contao verwendet wird, Sie dadurch womöglich die linke und Hauptspalte mit den pull-/push- Klassen verschieben müssen. ';
-$GLOBALS['TL_LANG']['tl_layout']['bootstrap_mainClass']['0']         = 'CSS-Grid Klassen der Hauptspalte';
-$GLOBALS['TL_LANG']['tl_layout']['bootstrap_mainClass']['1']         = 'Geben Sie die CSS-Grid Klassen der Hauptspalte an. Bedenken Sie, dass das Holygrail-Layout von Contao verwendet wird, Sie dadurch womöglich die linke und Hauptspalte mit den pull-/push- Klassen verschieben müssen. ';
-$GLOBALS['TL_LANG']['tl_layout']['bootstrap_rightClass']['0']        = 'CSS-Grid Klassen der rechten Spalte';
-$GLOBALS['TL_LANG']['tl_layout']['bootstrap_rightClass']['1']        = 'Geben Sie die CSS-Grid Klassen der rechten Spalte an.';
-$GLOBALS['TL_LANG']['tl_layout']['bootstrap_stylesheets']['0']       = 'Bootstrap Stylesheets';
-$GLOBALS['TL_LANG']['tl_layout']['bootstrap_stylesheets']['1']       = 'Bitte wählen Sie die gewünschten Dateien aus und konfigurieren sie. Sie werden beim Speichern zum Layout hinzugefügt.';
-$GLOBALS['TL_LANG']['tl_layout']['bs_asseticFilter']['0']            = 'Assetic-Filter';
-$GLOBALS['TL_LANG']['tl_layout']['bs_asseticFilter']['1']            = 'Bitte wählen Sie einen Assetic-Filter oder eine Filterkette aus, die für diese Datei benutzt werden soll.';
-$GLOBALS['TL_LANG']['tl_layout']['bs_conditional']['0']              = 'Conditional comment';
-$GLOBALS['TL_LANG']['tl_layout']['bs_conditional']['1']              = 'Conditional comments erlauben es, Dateien zu referenzieren, die sich ausschliesslich an den Internet Explorer richten. Nur die Bedingung für "lte IE 7" oder "IE 9" muss eingegeben werden. Der Prefix "&lt;!--[if" bzw. Suffix "]&gt;" wird automatisch hinzugefügt.';
-$GLOBALS['TL_LANG']['tl_layout']['bs_containerClass']['0']           = 'Container-Klasse';
-$GLOBALS['TL_LANG']['tl_layout']['bs_containerClass']['1']           = 'Bitte geben Sie die CSS-Klasse für den Container ein. Sie können <em>container</em>, <em>container-fluid</em> oder eine beliebige andere Klasse benutzen.';
-$GLOBALS['TL_LANG']['tl_layout']['bs_containerElement']['0']         = 'Container-Element';
-$GLOBALS['TL_LANG']['tl_layout']['bs_containerElement']['1']         = 'Bitte wählen Sie das Element aus, das als Bootstrap-Container dienen soll.';
-$GLOBALS['TL_LANG']['tl_layout']['bs_footerClass']['0']              = 'CSS-Klasse Footer';
-$GLOBALS['TL_LANG']['tl_layout']['bs_footerClass']['1']              = 'Kann beispielsweise benutzt werden um dem Footer die Klasse .container zuzuweisen.';
-$GLOBALS['TL_LANG']['tl_layout']['bs_headerClass']['0']              = 'CSS-Klasse Header';
-$GLOBALS['TL_LANG']['tl_layout']['bs_headerClass']['1']              = 'Kann beispielsweise benutzt werden um dem Header die Klasse .container zuzuweisen.';
-$GLOBALS['TL_LANG']['tl_layout']['bs_importFile']['0']               = 'Datei';
-$GLOBALS['TL_LANG']['tl_layout']['bs_importFile']['1']               = 'Wählen Sie eine verfügbare Datei aus.';
-$GLOBALS['TL_LANG']['tl_layout']['bs_importJavaScripts']['0']        = 'Bootstrap-JavaScript importieren';
-$GLOBALS['TL_LANG']['tl_layout']['bs_importJavaScripts']['1']        = 'Importiert die von Contao-Bootstrap mitgelieferten JavaScript-Dateien.';
-$GLOBALS['TL_LANG']['tl_layout']['bs_importStylesheets']['0']        = 'Bootstrap-CSS importieren';
-$GLOBALS['TL_LANG']['tl_layout']['bs_importStylesheets']['1']        = 'Importiert die von Contao-Bootstrap mitgelieferten Stylesheets.';
-$GLOBALS['TL_LANG']['tl_layout']['bs_javaScripts']['0']              = 'Bootstrap-JavaScript';
-$GLOBALS['TL_LANG']['tl_layout']['bs_javaScripts']['1']              = 'Bitte wählen Sie die gewünschten Dateien aus und konfigurieren sie. Beim Speichern werden sie dem Layout hinzugefügt.';
-$GLOBALS['TL_LANG']['tl_layout']['bs_leftClass']['0']                = 'CSS-Klasse linke Spalte';
-$GLOBALS['TL_LANG']['tl_layout']['bs_leftClass']['1']                = 'Setzen Sie CSS-Gridklassen für die rechte Spalte. Bitte beachten Sie, dass das Holygrail-Layout von Contao benutzt wird. Möglicherweise müssen Sie daher die linke und die Hauptspalte tauschen indem Sie .pull- und .push-Klassen benutzen.';
-$GLOBALS['TL_LANG']['tl_layout']['bs_mainClass']['0']                = 'CSS-Klasse Hauptspalte';
-$GLOBALS['TL_LANG']['tl_layout']['bs_mainClass']['1']                = 'Hiermit kann die Gridklasse für die Hauptspalte gesetzt werden. Bitte beachten Sie, dass das Holygrail-Layout von Contao benutzt wird. Möglicherweise müssen Sie daher die linke und die Hauptspalte tauschen indem Sie .pull- und .push-Klassen benutzen.';
-$GLOBALS['TL_LANG']['tl_layout']['bs_rightClass']['0']               = 'CSS-Klasse rechte Spalte';
-$GLOBALS['TL_LANG']['tl_layout']['bs_rightClass']['1']               = 'Setzen Sie CSS-Gridklassen für die rechte Spalte.';
-$GLOBALS['TL_LANG']['tl_layout']['bs_stylesheets']['0']              = 'Bootstrap-Stylesheets';
-$GLOBALS['TL_LANG']['tl_layout']['bs_stylesheets']['1']              = 'Bitte wählen Sie die gewünschten Dateien aus und konfigurieren sie. Beim Speichern werden sie dem Layout hinzugefügt.';
-$GLOBALS['TL_LANG']['tl_layout']['container_legend']                 = 'Container';
-$GLOBALS['TL_LANG']['tl_layout']['safeTpl']                          = 'Im Theme verfügbar';
-$GLOBALS['TL_LANG']['tl_layout']['unsafeTpl']                        = 'In anderen Themes';
-$GLOBALS['TL_LANG']['tl_layout']['viewport']['0']                    = 'Viewport';
-$GLOBALS['TL_LANG']['tl_layout']['viewport']['1']                    = 'Wählen Sie den Viewport für das Layout an. Dieser wird anders als Contao Standard permanent eingebunden. Desktopbrowser ignorieren diesen.';
-
+$GLOBALS['TL_LANG']['tl_layout']['bs_asseticFilter']['0']     = 'Assetic-Filter';
+$GLOBALS['TL_LANG']['tl_layout']['bs_asseticFilter']['1']     = 'Bitte wählen Sie einen Assetic-Filter oder eine Filterkette aus, die für diese Datei benutzt werden soll.';
+$GLOBALS['TL_LANG']['tl_layout']['bs_conditional']['0']       = 'Conditional comment';
+$GLOBALS['TL_LANG']['tl_layout']['bs_conditional']['1']       = 'Conditional comments erlauben es, Dateien zu referenzieren, die sich ausschliesslich an den Internet Explorer richten. Nur die Bedingung für "lte IE 7" oder "IE 9" muss eingegeben werden. Der Prefix "&lt;!--[if" bzw. Suffix "]&gt;" wird automatisch hinzugefügt.';
+$GLOBALS['TL_LANG']['tl_layout']['bs_containerClass']['0']    = 'Container-Klasse';
+$GLOBALS['TL_LANG']['tl_layout']['bs_containerClass']['1']    = 'Bitte geben Sie die CSS-Klasse für den Container ein. Sie können <em>container</em>, <em>container-fluid</em> oder eine beliebige andere Klasse benutzen.';
+$GLOBALS['TL_LANG']['tl_layout']['bs_containerElement']['0']  = 'Container-Element';
+$GLOBALS['TL_LANG']['tl_layout']['bs_containerElement']['1']  = 'Bitte wählen Sie das Element aus, das als Bootstrap-Container dienen soll.';
+$GLOBALS['TL_LANG']['tl_layout']['bs_footerClass']['0']       = 'CSS-Klasse Footer';
+$GLOBALS['TL_LANG']['tl_layout']['bs_footerClass']['1']       = 'Kann beispielsweise benutzt werden um dem Footer die Klasse .container zuzuweisen.';
+$GLOBALS['TL_LANG']['tl_layout']['bs_headerClass']['0']       = 'CSS-Klasse Header';
+$GLOBALS['TL_LANG']['tl_layout']['bs_headerClass']['1']       = 'Kann beispielsweise benutzt werden um dem Header die Klasse .container zuzuweisen.';
+$GLOBALS['TL_LANG']['tl_layout']['bs_importFile']['0']        = 'Datei';
+$GLOBALS['TL_LANG']['tl_layout']['bs_importFile']['1']        = 'Wählen Sie eine verfügbare Datei aus.';
+$GLOBALS['TL_LANG']['tl_layout']['bs_importJavaScripts']['0'] = 'Bootstrap-JavaScript importieren';
+$GLOBALS['TL_LANG']['tl_layout']['bs_importJavaScripts']['1'] = 'Importiert die von Contao-Bootstrap mitgelieferten JavaScript-Dateien.';
+$GLOBALS['TL_LANG']['tl_layout']['bs_importStylesheets']['0'] = 'Bootstrap-CSS importieren';
+$GLOBALS['TL_LANG']['tl_layout']['bs_importStylesheets']['1'] = 'Importiert die von Contao-Bootstrap mitgelieferten Stylesheets.';
+$GLOBALS['TL_LANG']['tl_layout']['bs_javaScripts']['0']       = 'Bootstrap-JavaScript';
+$GLOBALS['TL_LANG']['tl_layout']['bs_javaScripts']['1']       = 'Bitte wählen Sie die gewünschten Dateien aus und konfigurieren sie. Beim Speichern werden sie dem Layout hinzugefügt.';
+$GLOBALS['TL_LANG']['tl_layout']['bs_leftClass']['0']         = 'CSS-Klasse linke Spalte';
+$GLOBALS['TL_LANG']['tl_layout']['bs_leftClass']['1']         = 'Setzen Sie CSS-Gridklassen für die rechte Spalte. Bitte beachten Sie, dass das Holygrail-Layout von Contao benutzt wird. Möglicherweise müssen Sie daher die linke und die Hauptspalte tauschen indem Sie .pull- und .push-Klassen benutzen.';
+$GLOBALS['TL_LANG']['tl_layout']['bs_mainClass']['0']         = 'CSS-Klasse Hauptspalte';
+$GLOBALS['TL_LANG']['tl_layout']['bs_mainClass']['1']         = 'Hiermit kann die Gridklasse für die Hauptspalte gesetzt werden. Bitte beachten Sie, dass das Holygrail-Layout von Contao benutzt wird. Möglicherweise müssen Sie daher die linke und die Hauptspalte tauschen indem Sie .pull- und .push-Klassen benutzen.';
+$GLOBALS['TL_LANG']['tl_layout']['bs_rightClass']['0']        = 'CSS-Klasse rechte Spalte';
+$GLOBALS['TL_LANG']['tl_layout']['bs_rightClass']['1']        = 'Setzen Sie CSS-Gridklassen für die rechte Spalte.';
+$GLOBALS['TL_LANG']['tl_layout']['bs_stylesheets']['0']       = 'Bootstrap-Stylesheets';
+$GLOBALS['TL_LANG']['tl_layout']['bs_stylesheets']['1']       = 'Bitte wählen Sie die gewünschten Dateien aus und konfigurieren sie. Beim Speichern werden sie dem Layout hinzugefügt.';
+$GLOBALS['TL_LANG']['tl_layout']['container_legend']          = 'Container';
+$GLOBALS['TL_LANG']['tl_layout']['safeTpl']                   = 'Im Theme verfügbar';
+$GLOBALS['TL_LANG']['tl_layout']['unsafeTpl']                 = 'In anderen Themes';
+$GLOBALS['TL_LANG']['tl_layout']['viewport']['0']             = 'Viewport';
+$GLOBALS['TL_LANG']['tl_layout']['viewport']['1']             = 'Wählen Sie den Viewport für das Layout an. Dieser wird anders als Contao Standard permanent eingebunden. Desktopbrowser ignorieren diesen.';